requirements.txt
Django==3.1.6
gunicorn==20.0.4
tensorflow-cpu==2.4.1
tensorflow-hub
tensorflow-text
